requirements.txt
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
geoip2==2.9.0
idna==2.8
ipython-genutils==0.2.0
jsonschema==3.0.1
jupyter-core==4.4.0
maxminddb==1.4.1
nbformat==4.4.0
pandas==0.24.2
plotly==3.9.0
progressbar2==3.39.3
pyrsistent==0.15.2
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.1
requests==2.22.0
retrying==1.3.3
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
